# DTO/InputChat.cs
namespace WEBAPI_m1IL_1.DTO
{
    public class InputChat
    {
        public string Prompt { get; set; }
        public string? Image { get; set; }
        public string Context { get; set; }
        public string? Model { get; set; }
    }
}